+++ xcm/xcm-builder/src/barriers.rs
@@ -16,21 +16,19 @@
 
 //! Various implementations for `ShouldExecute`.
 
+use crate::{CreateMatcher, MatchXcm};
 use frame_support::{
 	ensure,
 	traits::{Contains, Get},
 };
 use polkadot_parachain::primitives::IsSystem;
 use sp_std::{cell::Cell, marker::PhantomData, ops::ControlFlow, result::Result};
-use xcm::{
-	latest::{
-		Instruction::{self, *},
-		InteriorMultiLocation, Junction, Junctions,
-		Junctions::X1,
-		MultiLocation, Weight,
-		WeightLimit::*,
-	},
-	CreateMatcher, MatchXcm,
+use xcm::latest::{
+	Instruction::{self, *},
+	InteriorMultiLocation, Junction, Junctions,
+	Junctions::X1,
+	MultiLocation, Weight,
+	WeightLimit::*,
 };
 use xcm_executor::traits::{OnResponse, ShouldExecute};
