+++ xcm/xcm-builder/src/lib.rs
@@ -76,6 +76,9 @@ pub use weight::{
 mod matches_token;
 pub use matches_token::{IsAbstract, IsConcrete};
 
+mod matcher;
+pub use matcher::{CreateMatcher, MatchXcm, Matcher};
+
 mod filter_asset_location;
 pub use filter_asset_location::{Case, NativeAsset};
